Updated from .env to database

+++ src/buttons/ignore-user.js
@@ -1,5 +1,8 @@
-const { Interaction, Client, EmbedBuilder } = require("discord.js");
-const { env } = require("../env");
+const {
+    CommandInteraction,
+    Client,
+    EmbedBuilder
+} = require("discord.js");
 
 module.exports = {
     data: {
@@ -7,7 +10,7 @@ module.exports = {
     },
     /**
     * 
-    * @param {Interaction} interaction
+    * @param {CommandInteraction} interaction
     * @param {Client} client
     */
     async execute(interaction, client) {
@@ -16,29 +19,34 @@ module.exports = {
         var member = await interaction.guild.members.fetch(memberId);
 
         try {
+            const guild = await client.database.db("kiwi").collection("guilds").findOne(
+                { guildId: interaction.guildId }
+            );
+
+            if (!guild) return interaction.followUp("This guild is not in the database.");
             await member.send(`You have been **denied** from **${interaction.guild.name}**`)
+            await member.kick("Denied");
+            await interaction.message.delete();
+
+            if (guild.logsChannel) {
+
+                var log = await interaction.guild.channels.cache.get(guild.logsChannel);
+                var em = new EmbedBuilder()
+                    .setTitle(member.user.username + "#" + member.user.discriminator)
+                    .setThumbnail(member.user.avatarURL())
+                    .addFields(
+                        { name: "Mention", value: `<@${member.user.id}>` },
+                        { name: "Denied By", value: `<@${interaction.member.id}>` },
+                        { name: "Action", value: "Kicked" }
+                    )
+                    .setColor(0xFF474D)
+
+                await log.send({
+                    embeds: [em]
+                });
+            }
         } catch (e) {
             console.log(e);
-        }
-        await member.kick("Denied");
-        await interaction.message.delete();
-
-        if (env.LOGS_CHANNEL) {
-
-            var log = await interaction.guild.channels.cache.get(env.LOGS_CHANNEL);
-            var em = new EmbedBuilder()
-                .setTitle(member.user.username + "#" + member.user.discriminator)
-                .setThumbnail(member.user.avatarURL())
-                .addFields(
-                    { name: "Mention", value: `<@${member.user.id}>` },
-                    { name: "Denied By", value: `<@${interaction.member.id}>` },
-                    { name: "Action", value: "Kicked" }
-                )
-                .setColor(0xFF474D)
-
-            await log.send({
-                embeds: [em]
-            });
         }
 
     }
